Fixing Initialization of Neurons
In basespikingchannel.cc, if a unit was present in the xml file, it was
initialized with default values instead of the values of the spiking
channel.
Now the behaviour changes, such that the properties of the
spikingchannel is propagated to all units that are children of this xml
node.

Also the parameter for square and radial sampling schemes to add jitter
to the position of neurons is not processed.
The parameter "noise_pos" can now be used as an attribute of rdSampling
and sqSampling.